### 2023-09-20 #5859 @editorialbot
[PRE REVIEW]: GECo: A collection of solvers for the self-gravitating Vlasov equations
**Submitting author:** @alogg (<a href="http://orcid.org/0000-0002-1547-4773">Anders Logg</a>)
**Repository:** https://github.com/anderslogg/geco
**Branch with paper.md** (empty if default branch): paper-joss-2023
**Version:** 1.0.0
