### 2023-10-16 #5859 @dfm
@elleryames, @alogg — In the meantime it looks like most of the references in the compiled document are not getting resolved. Can you check your BibTeX and try to work out what's going on? There are some methods [outlined on the JOSS docs page](https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles) to check the compilation. Thanks!
