### 2023-10-16 #5859 @anderslogg
@dfm We have followed the instructions on the web page very carefully but still the figure references are broken.

The instructions on the web page look like this:

    ![Caption for example figure.\label{fig:example}](figure.png)
    and referenced from text using \autoref{fig:example}.

And we have the following in our paper:

    \autoref{fig:Solution}

    ![Torus spatial density on quarter plane computational domain \label{fig:Solution}](./figures/density_computational_domain.png)

We have also built the document using the Docker instructions as listed in the instruction but the references break for some reason.
