### 2023-10-16 #5859 @anderslogg
@dfm Thanks! Will fix. So it was just a matter of separating the figures with newlines?
