### 2023-10-16 0a65e7f
Add figure captions
