### 2023-10-23 #5979 @editorialbot
[REVIEW]: GECo: A collection of solvers for the self-gravitating Vlasov equations
**Submitting author:** @anderslogg (<a href="http://orcid.org/0000-0002-1547-4773">Anders Logg</a>)
**Repository:** https://github.com/anderslogg/geco
**Branch with paper.md** (empty if default branch): paper-joss-2023
**Version:** 1.0.1
