### 2023-10-28 #5979 @pmocz
> hi @anderslogg, I am trying to run the demos using the python3 ev_torus.py command in documentation, but it seems these require a module named dolfin.
>
> I would find a list of dependencies to install helpful in the installation page. Especially if there are multiple, it's good to provide an installation file to run to sort these out (especially if there are different versions floating around). It's very possible I am just having trouble installing this as I get an error:
>
> ERROR: Could not find a version that satisfies the requirement dolfin (from versions: none) ERROR: No matching distribution found for dolfin

I also ran into the same issue and had to run the docker version
