### 2023-10-30 #5979 @anderslogg
@michellegurevich @pmocz This is expected and also stated in the installation instructions here:

https://github.com/anderslogg/geco/blob/main/docs/installation.md

The user needs to install FEniCS version 2019.2.0 (which includes the package DOLFIN that you get an error message about). However, as also stated in the docs, it's much easier to use the provided Docker image that has all the dependencies installed.
