### 2023-12-09 #5979 @elleryames
Thanks @dfm  and @anderslogg for getting this through, and thanks @pmocz  and @michellegurevich for the reviews!

I have read through the paper, updated my affiliation and fixed punctuation in a couple places. Looks good to me!
