### 2023-12-11 #5979 @anderslogg
@dfm We have now proofread, fixed a few minor typos, updated affiliations, incremented the version number to 1.0.1 and archived the software at https://zenodo.org/records/10351519.
